requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-multipart
langchain
langchain-google-genai
langchain-groq
langchain-huggingface
chromadb
faiss-cpu
python-dotenv
python-docx
PyPDF2
pandas
openpyxl
numpy
streamlit
graphviz
